==> 00_run_all.R <==
source("03_positions_and_org.R")
source("04_employee_info.R")
source("05_initialize_employee_desk_history.R")
rm(list=ls())
source("06_grow_employee_desk_history.R")
rm(list=ls())
source("07_create_performance_reviews_salaries.R")
rm(list=ls())
source("08_employee_info_cleanup.R")
rm(list=ls())
source("09_recruiting.R")
rm(list=ls())
source("10_views.R")
rm(list=ls())
source("11_contact_info.R")
rm(list=ls())
source("12_education_skills.R")
